requirements.txt
certifi==2022.6.15
charset-normalizer==2.0.12
idna==3.3
python-decouple==3.6
pytz==2021.1
regex==2021.4.4
requests==2.26.0
ticktick-py==2.0.2
urllib3==1.26.7
